refactor(tooling): split lsp smoke support
